chore: refactor to use proper dev pipeline
