+++ packages/astro/src/core/app/index.ts
@@ -1,6 +1,8 @@
 export { App } from './app.js';
 export { BaseApp, type RenderErrorOptions, type RenderOptions } from './base.js';
 export { deserializeManifest, fromRoutingStrategy, toRoutingStrategy } from './common.js';
+export { DevApp } from './dev/app.js';
+export { createConsoleLogger } from './logging.js';
 export {
 	deserializeRouteData,
 	deserializeRouteInfo,
